Initial conversion to C++
