+++ test/failed_spares/CMakeLists.txt
@@ -9,7 +9,7 @@
 #
 
 add_executable(fenix_failed_spares fenix_failed_spares.c)
-target_link_libraries(fenix_failed_spares fenix MPI::MPI_C)
+target_link_libraries(fenix_failed_spares fenix MPI::MPI_CXX)
 
 add_test(NAME failed_spares 
   COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} 6 ${MPIEXEC_PREFLAGS} fenix_failed_spares ${MPIEXEC_POSTFLAGS} 3 1 3 4 )
