+++ test/message_replay/CMakeLists.txt
@@ -8,7 +8,7 @@
 #  directory.
 #
 
-add_executable(fenix_no_jump_test fenix_no_jump_test.c)
-target_link_libraries(fenix_no_jump_test fenix MPI::MPI_CXX)
+add_executable(fenix_message_logging_test fenix_message_logging_test.cxx)
+target_link_libraries(fenix_message_logging_test fenix MPI::MPI_CXX)
 
-add_test(NAME no_jump COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} 5 ${MPIEXEC_PREFLAGS} fenix_no_jump_test ${MPIEXEC_POSTFLAGS} "1")
+add_test(NAME request_cancelled COMMAND ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} 5 ${MPIEXEC_PREFLAGS} fenix_message_logging_test ${MPIEXEC_POSTFLAGS} "1")
